+++ NEWS.md
@@ -1,3 +1,7 @@
+# seeker 1.0.3
+* Made `getProbeGeneMapping()` test check for version number of mapping package.
+* Revised arguments for `seekerArray()` and updated tests and vignettes accordingly.
+
 # seeker 1.0.2
 * Set default parent directory for `seekerArray()`.
 * Added custom CDF support for platform GPL17400.